~checks for null and empty fields,+some logic for checking new bid's amount against old bid

+++ src/api.js
@@ -27,20 +27,19 @@ const secrets = {
   infuraSecret: process.env.INFURA_API_SECRET,
 };
 
-// Function to check null and empty auction creation fields
-function checkNullCreateFields(...args) {
-  let createFields = [
-    "account",
-    "tokenID",
+// Function to check null and empty encode bid fields
+function checkNullEncodeFields(...args) {
+  let encodeFields = [
+    "auctionId",
+    "zAuctionAddress",
+    "chainId",
+    "bidAmt",
     "contractAddress",
-    "startTime",
-    "endTime",
-    "minBid",
-    "auctionType",
+    "tokenId"
   ];
   for (let i = 0; i < args.length; i++) {
     if (args[i] == null || !/\S/.test(args[i])) {
-      return { data: createFields[i], value: false };
+      return { data: encodeFields[i], value: false };
     }
   }
   return { data: null, value: true };
@@ -48,7 +47,7 @@ function checkNullCreateFields(...args) {
 
 // Function to check null and empty bid request fields
 function checkNullBidFields(...args) {
-  let bidFields = ["account", "bidAmt", "bidMsg"];
+  let bidFields = ["seller", "account", "tokenId", "contractAddress", "bidAmt", "bidMsg"];
   for (let i = 0; i < args.length; i++) {
     if (args[i] == null || !/\S/.test(args[i])) {
       return { data: bidFields[i], value: false };
@@ -57,11 +56,33 @@ function checkNullBidFields(...args) {
   return { data: null, value: true };
 }
 
+// Function to check null and empty current bid request fields
+function checkNullCurrentBidFields(...args) {
+  let currentBidFields = ["seller", "contractAddress", "tokenId"];
+  for (let i = 0; i < args.length; i++) {
+    if (args[i] == null || !/\S/.test(args[i])) {
+      return { data: currentBidFields[i], value: false };
+    }
+  }
+  return { data: null, value: true };
+}
+
 // Returns a encoded data to be signed
-router.get("/encodebid", async (req, res, next) => {
+router.get("/encodeBid", async (req, res, next) => {
   try {
+    let result = checkNullEncodeFields(
+      req.body.auctionId,
+      req.body.zAuctionAddress,
+      req.body.chainId,
+      req.body.bidAmt,
+      req.body.contractAddress,
+      req.body.tokenId
+    );
+    if (result["value"] == false) {
+      return res.status(400).send({ message: result["data"] + " not found" });
+    }
     let params = ethers.utils.defaultAbiCoder(['uint256','address','uint8','uint256', 'address', 'uint256', 'uint256','uint256','uint256'],
-    [req.body.auctionID, req.body.zauctionAddress, req.body.chainId, req.body.bidAmt, req.body.contractAddress, req.body.tokenID, 0, 0, 9999999999999]);      
+    [req.body.auctionId, req.body.zAuctionAddress, req.body.chainId, req.body.bidAmt, req.body.contractAddress, req.body.tokenId, 0, 0, 9999999999999]);      
     return ethers.utils.keccak256(params);
   } catch (error) {
     next(error);
@@ -74,7 +95,7 @@ router.post("/bid", async (req, res, next) => {
     let result = checkNullBidFields(
       req.body.seller,
       req.body.account, // account of the bidder
-      req.body.tokenID,
+      req.body.tokenId,
       req.body.contractAddress,
       req.body.bidAmt,
       req.body.bidMsg // signed msg
@@ -83,11 +104,11 @@ router.post("/bid", async (req, res, next) => {
       return res.status(400).send({ message: result["data"] + " not found" });
     }
     // generate auctionId
-    let idString = req.body.contractAddress + req.body.tokenID + req.body.seller;
+    let idString = req.body.contractAddress + req.body.tokenId + req.body.seller;
     let idStringBytes = ethers.utils.toUtf8Bytes(idString);
     let auctionId = ethers.utils.keccak256(idStringBytes);
     //console.log("New auctionId is", auctionId);
-    // pull auction from fleek
+    // pull auction from fleek with given auctionId
     await fleek
       .get({
         apiKey: secrets.apiKey,
@@ -101,35 +122,41 @@ router.post("/bid", async (req, res, next) => {
           const data = {
             seller: oldAuction.seller,
             account: oldAuction.account,
-            tokenID: oldAuction.tokenID,
+            tokenId: oldAuction.tokenId,
             contractAddress: oldAuction.contractAddress,
             currentBidder: req.body.account,
             currentBid: req.body.bidAmt,
             bidMsg: req.body.bidMsg,
           };
-          // delete the old auction
-          await fleek
-            .deleteFile({
-              apiKey: secrets.apiKey,
-              apiSecret: secrets.apiSecret,
-              key: auctionID,
-            })
-            .then(async () => {
-              // and upload new auction under the same name (key)
-              await fleek
-                .upload({
-                  apiKey: secrets.apiKey,
-                  apiSecret: secrets.apiSecret,
-                  key: auctionID,
-                  data: JSON.stringify(data),
-                })
-                .then(() => res.status(200).send({ message: "Ok" }));
-            });
+          // if the new bid amount is greater than the old bid amount
+          if (req.body.bidAmt > oldAuction.bidAmt) {
+            // delete the old auction
+            await fleek
+              .deleteFile({
+                apiKey: secrets.apiKey,
+                apiSecret: secrets.apiSecret,
+                key: auctionId,
+              })
+              .then(async () => {
+                // and upload new auction under the same name (key)
+                await fleek
+                  .upload({
+                    apiKey: secrets.apiKey,
+                    apiSecret: secrets.apiSecret,
+                    key: auctionId,
+                    data: JSON.stringify(data),
+                  })
+                  .then(() => res.status(200).send({ message: "Ok" }));
+              });
+
+          } else {
+            res.status(400).send({ message: "New bid amount must be greater than current bid amount"});
+          }
         } else {
           const data = {
             seller: req.body.seller,
             account: req.body.account,
-            tokenID: req.body.tokenID,
+            tokenId: req.body.tokenId,
             contractAddress: req.body.contractAddress,
             currentBidder: req.body.account,
             currentBid: req.body.bidMsg,
@@ -138,7 +165,7 @@ router.post("/bid", async (req, res, next) => {
                 .upload({
                   apiKey: secrets.apiKey,
                   apiSecret: secrets.apiSecret,
-                  key: auctionID,
+                  key: auctionId,
                   data: JSON.stringify(data),
                 })
                 .then(() => res.status(200).send({ message: "Ok" }));
@@ -152,14 +179,16 @@ router.post("/bid", async (req, res, next) => {
 // Endpoint to return current highest bid given seller, contract address, and tokenid
 router.get("/currentBid", limiter, async (req, res, next) => {
   try {
-    if (req.params.seller == null || req.params.contractAddress == null || req.params.tokenID == null) {
-      return res.send({
-        status: "false",
-        message: "Please provide an account, contract address, and token id",
-      });
+    let result = checkNullCurrentBidFields(
+      req.body.seller,
+      req.body.contractAddress,
+      req.body.tokenId
+    );
+    if (result["value"] == false) {
+      return res.status(400).send({ message: result["data"] + " not found" });
     }
     // generate auctionId
-    let idString = req.body.contractAddress + req.body.tokenID + req.body.seller;
+    let idString = req.body.contractAddress + req.body.tokenId + req.body.seller;
     let idStringBytes = ethers.utils.toUtf8Bytes(idString);
     let auctionId = ethers.utils.keccak256(idStringBytes);
     // get file with key from fleek
@@ -167,7 +196,7 @@ router.get("/currentBid", limiter, async (req, res, next) => {
       .get({
         apiKey: secrets.apiKey,
         apiSecret: secrets.apiSecret,
-        key: auctionID,
+        key: auctionId,
       })
       .then((file) => {
         // parse file and return only bids
